Fix rendering of markers added before audio is loaded
The location of a marker for a specific time in the audio cannot be
correctly calculated until we know the duration of the audio. So any
markers added before the audio has been downloaded and processed need to
have their locations (aka "percentages") recalculated once the audio
duration is available.

This resolves issue #16.