The move and resize events should hold the whole position difference.
If the resize event does not hold the whole position difference, there will be a lag between the position of the cursor and the position of the region as the user drags around. This is because the previous code added pixels incrementally.

For the move event, the issue isn't as evident, but it is when the mouse moves off the wavesurfer object with the mouse still in the "down" position.