Remove pointer events from the left and right sides of the wave and change zIndex to 3.
The reasoning behind this is to allow, for example, the region plugin to have a styling such that the regions are below the waves (so that the waves look clearer and not blended with the regions), while still allowing resize events on the regions. Otherwise, the left/right sections of the wave block the resize elements of the regions, since their zIndex needs to be necessarily higher to be drawn over the regions.