Add getters and setters for height and color options (#1145)
Allow updating the `waveColor`, `progressColor`, `cursorColor` and
`height` options that were passed as part of the options for create.

For the wave color options, simply ensure that the wave is
redrawn. For the height option, delegate to `drawer.setHeight` to
handle the update. Note that - while drawer applies the `pixelRatio`
itself when initializing from `params` - `setHeight` expects a height
that already has been multiplied by `pixelRatio`.

The cursor is a border on one of the wave elements. Add an
`updateCursor` method to the drawer interface to allow updating the
cursor related styles.