Ensure Travis build fails when webpack build fails
Commands in package.json scrpts were chained using "&". This causes
parallel execution which might be desired, but causes exit code of all
but last command to be ignored. Use "&&" operator instead to execute
commands sequentially and fail as soon as one does not succeed.

In particular, this ensures the build fails when prettier finds
problems.