Set isReady attribute to false on empty (#1396)
The isReady attribute is set to false at the creation of a wavesurfer. It it set to true when a song is loaded but it is never set to false when emptying the wavesufer. As a result, after the first song is loaded, isReady return true even if the wavesurfer is empty.