regions plugin: display regions before file load (#1441)
Pre-generated peaks waits for the audio length, either from loaded audio or from explicitDuration. This PR adds explicitDuration as an argument for the WebAudio constructor. This allows it to be defined up front.

I've also updated the regions plugin initialization to make use of this information so the "backend-created" event is used to add regions, rather than the "ready" event which only fires when a song has finished loading.

* Fix to make pre-generated peaks and regions to load together with an explicit duration

* Return empty peaks if envelope not set and buffer not set

* Avoid changing region limits based on duration because duration value changes as data is loaded

* address review comment: use camel-case

* remove listener in destroy

* rename duration option

* use duration param