Calculates the duration at event time
Previously the duration was being calculated once upon `bindEvents()`, but predefined regions may be being created before the duration is being set (and running `addRegion()` on the `ready` method of the WaveSurfer object doesn't create the DOM element). This meant that predefined regions couldn't be dragged or resized, since the duration was always 0.