avoid call to HTMLMediaElement#load when peaks are given (#1969)
* avoid call to HTMLMediaElement#load when peaks are given

calling `.load()` explicitly on a mediaelement entirely defeats the
"preload" setting and makes it impossible to avoid downloading
data before playback.

When HTMLMediaElement#load is called, all of the major browsers make
some request for the file.  Chrome and FF make a range request for the
entire file, but only report downloading them partially.  Safari reports
downloading the file in its entirety.

* code review; only avoid .load call when preload == 'none'

* add changelog entry

* bump changelog entry to next release

* add next section to changelog

Co-authored-by: Thijs Triemstra <info@collab.nl>