cursor plugin: fix element removal in destroy (#2462)
* Change how elements are removed on destroying

When the plugin destroy() method was called, I got an error because removeChild was called with something that wasn't a Node, for this.showTime and this.cursor. 
It seems that these members are Proxy objects, and the removeChild doesn't swallow that. Changing from removeChild to remove() on the Proxy objects seems to fix this.

* Update CHANGES.md