Improvements to rendering.
Adds a (local) constant that controls a number of bars to hide from the
start and end of the wave - makes it look better under the overlay

Changes rendering to be locked to starting at a multiple of the step.
E.g. if the bar is 1px and the gap is 1px, the bar step is 2. Before,
when we drag the start handle, we are alternating between starting the
rendering of bars at an odd number or an even number, which recalculates
and subtly shimmers the bars. Now we will always start rendering at even
numbers, so the recalculation will not happen